+++ src/lowpassfilter.h
@@ -32,7 +32,7 @@ class LowPassFilterPrivate;
 
 class LowPassFilter {
 public:
-    LowPassFilter(unsigned int order, unsigned int frameRate, double cornerFrequency, unsigned int fftFrameSize);
+    LowPassFilter(unsigned int order, unsigned int frameRate, float cornerFrequency, unsigned int fftFrameSize);
     ~LowPassFilter();
     void filter(AudioData& audio, Workspace& workspace, unsigned int shortcutFactor = 1) const;
     [[nodiscard]] auto getCoefficients() const -> void const*; // for unit testing only
